% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_columns.R
\name{check_columns}
\alias{check_columns}
\title{Check columns of data frame}
\usage{
check_columns(spocc_name, fields)
}
\arguments{
\item{spocc_name}{A list of columns from the query.}

\item{fields}{A list of fields.}
}
\value{
Returns a data frame with desired columns.
}
\description{
The \code{check_columns()} function provides the columns requested.
This function requires no additional packages.
}
\keyword{internal}
