% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_class.R
\name{correct_class}
\alias{correct_class}
\title{gatoRs Download - Correct classes of data frame columns}
\usage{
correct_class(
  df,
  scientific.name = "scientificName",
  genus = "genus",
  species = "specificEpithet",
  infraspecific.epithet = "infraspecificEpithet",
  id = "ID",
  occ.id = "occurrenceID",
  basis.of.record = "basisOfRecord",
  event.date = "eventDate",
  year = "year",
  month = "month",
  day = "day",
  col.code = "collectionCode",
  recorded.by = "recordedBy",
  country = "country",
  county = "county",
  state = "stateProvince",
  locality = "locality",
  latitude = "latitude",
  longitude = "longitude",
  coord.uncertainty = "coordinateUncertaintyInMeters",
  info.withheld = "informationWithheld",
  habitat = "habitat",
  aggregator = "aggregator"
)
}
\arguments{
\item{df}{Data frame returned by \code{gator_download()}.}

\item{scientific.name}{Default = "scientificName". The name of the scientific name column in the data frame.}

\item{genus}{Default = "genus". The name of the genus column in the data frame.}

\item{species}{Default = "specificEpithet". The name of the specific epithet column in the data frame.}

\item{infraspecific.epithet}{Default = "infraspecificEpithet". The name of the infraspecific epithet column in the data frame.}

\item{id}{Default = "ID". The name of the id column in the data frame, which contains unique IDs defined from GBIF or iDigBio.}

\item{occ.id}{Default = "occurrenceId". The name of the occurrence ID column in the data frame.}

\item{basis.of.record}{Default = "basisOfRecord". The name of the basis of record column in the data frame.}

\item{event.date}{Default = "eventDate". The name of the event date column in the data frame.}

\item{year}{Default = "year". The name of the event date year column in the data frame.}

\item{month}{Default = "month". The name of the event date month column in the data frame.}

\item{day}{Default = "day". The name of the event date day column in the data frame.}

\item{col.code}{Default = "collectionCode". The name of the collection code column in the data frame.}

\item{recorded.by}{Default = "recordedBy". The name of the collector name column in the data frame.}

\item{country}{Default = "country". The name of the country column in the data frame.}

\item{county}{Default = "county". The name of the county column in the data frame.}

\item{state}{Default = "stateProvince". The name of the state/province column in the data frame.}

\item{locality}{Default = "locality". The name of the locality column in the data frame.}

\item{latitude}{Default = "latitude". The name of the latitude column in the data frame.}

\item{longitude}{Default = "longitude". The name of the longitude column in the data frame.}

\item{coord.uncertainty}{Default = "coordinateUncertaintyInMeters". The name of the coordinate uncertainty column in the data frame.}

\item{info.withheld}{Default = "informationWithheld". The name of the information withheld column in the data frame.}

\item{habitat}{Default = "habitat". The name of the habitat column in the data frame.}

\item{aggregator}{Default = "aggregator". The name of the column in the data frame that identifies the aggregator
that provided the record.}
}
\value{
Returns data frame with corrected classes of each column.
}
\description{
The \code{correct_class()} function corrects the classes of each column in a data frame of your queried species.
This function requires no additional packages.
}
\details{
"aggregator", "infraspecific.epithet", and "recorded.by" columns will be skipped if they do not exist.
}
\keyword{internal}
