% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Downloaded data from gators_download() for Galax urceolata with default settings:
data <- gators_download(synonyms_list = c("Galax urceolata", "Galax aphylla"))}
\format{
A data frame with 6885 rows and 19 columns
}
\usage{
data
}
\description{
Downloaded data from gators_download() for Galax urceolata with default settings:
data <- gators_download(synonyms_list = c("Galax urceolata", "Galax aphylla"))
}
\keyword{internal}
