% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fix_names.R
\name{filter_fix_names}
\alias{filter_fix_names}
\title{filter_fix_names

This function filters a data frame for relevant results, based on the scientific name given.
This function requires packages base and dplyr.}
\usage{
filter_fix_names(df, synonyms_list, filter, accepted_name)
}
\arguments{
\item{df}{is a data frame with name column to be fixed}

\item{synonyms_list}{is a list of synonyms for a species}

\item{filter}{is the type of filter to be used--either "exact" or "fuzzy"}

\item{accepted_name}{is the accepted scientific name for the species}
}
\value{
returns df with filtered results
}
\description{
filter_fix_names

This function filters a data frame for relevant results, based on the scientific name given.
This function requires packages base and dplyr.
}
