% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_flagged.R
\name{find_flagged}
\alias{find_flagged}
\title{find_flagged}
\usage{
find_flagged(df)
}
\arguments{
\item{df}{is a dataframe of occurrence records.}
}
\value{
Return a copy of the original dataframe provided as input but without any points the user chose to remove.
}
\description{
This function allows you to find and map possible problematic points and remove these points, if desired.
}
\details{
This function requires packages dplyr, CoordinateCleaner, base, leaflet, magrittr. This function requires user input.
}
