% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_columns.R
\name{fix_columns}
\alias{fix_columns}
\title{fix_columns}
\usage{
fix_columns(df)
}
\arguments{
\item{df}{is a dataframe of occurrence records}
}
\value{
df is the original dataframe with the specified columns
}
\description{
This function selects columns to keep and corrects the class of each.
This function requires packages dplyr, base, lubridate, magrittr, stringr.
This function uses the correct_class function.
}
