% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_names.R
\name{fix_names}
\alias{fix_names}
\title{fix_names

This function fixes the capitalization of species names.
This function requires packages base and stringr. This function does not require user input.}
\usage{
fix_names(df)
}
\arguments{
\item{df}{is a dataframe with name column to be fixed}
}
\value{
returns df with fixed capitalization in name column
}
\description{
fix_names

This function fixes the capitalization of species names.
This function requires packages base and stringr. This function does not require user input.
}
