% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixAfterPeriod.R
\name{fixAfterPeriod}
\alias{fixAfterPeriod}
\title{fixAfterPeriod}
\usage{
fixAfterPeriod(substring)
}
\arguments{
\item{substring}{A substring from the name column of the data frame to be fixed.}
}
\value{
Returns the substring with fixed capitalization.
}
\description{
This function fixes capitalization of species names when there are periods in the name.
}
\details{
Requires packages base and stringr. Does not require user input.
}
