% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_idigbio.R
\name{get_idigbio}
\alias{get_idigbio}
\title{Used in gators_download() - Download data from Integrated Digitized Biocollections}
\usage{
get_idigbio(synonyms.list, limit = 1e+05)
}
\arguments{
\item{synonyms.list}{A list of affiliated names for your query.}

\item{limit}{Default = 100,000 (maximum). Set limit to the number of records requested for each element in synonyms.list.}
}
\value{
A data frame with desired columns from iDigBio.
}
\description{
The \code{get_idigbio()} function queries iDigBio for your desired species.
Limited to 100,000 record downloads.
}
\details{
This function uses the \code{correct_class()} function.
This function requires the packages ridigbio, magrittr, and dplyr.
}
\examples{
df <- get_idigbio(c("Galax urceolata", "Galax aphylla"), limit = 100)

}
