% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_idigbio_arc.R
\name{get_idigbio_arc}
\alias{get_idigbio_arc}
\title{Used in gators_download() - Download data from Integrated Digitized Biocollections that are not on GBIF}
\usage{
get_idigbio_arc(synonyms.list, idigbio.match = "fuzzy", fuzzy.ratio = 50)
}
\arguments{
\item{synonyms.list}{A list of affiliated names for your query.}

\item{idigbio.match}{Default = "fuzzy". Options are "fuzzy" or "exact".
The fuzzy match uses \href{https://github.com/rapidfuzz/RapidFuzz}{rapidfuzz} to calculate
a rapidfuzz_partial_token_set_ratio or similarity ratio between the two strings. This option
is helpful when word order differs or our strings are partial matches.}

\item{fuzzy.ratio}{Default = 50\%. We let more match than probably are helpful, and return all ratios greater than 50\%.}
}
\value{
A data frame with desired columns from iDigBio.
}
\description{
The \code{get_idigbio_arc()} function queries an archive of iDigBio that contains only
recordsets that were not uploaded to GBIF by June 2026. Since funding ended for iDigBio,
they are unable to host data. Though the vast majority of recordsets migrated to GBIF,
we wanted to make sure we had information available for the remaining.
}
\details{
This function uses the \code{correct_class()} function and \code{setupduckDB()}.
This function requires the packages magrittr, dplyr, duckDB, arrow, and DBI.
}
\examples{
if(exists("crazy")){
df <- get_idigbio_arc("Galax")
}
}
