% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbien.R
\name{getbien}
\alias{getbien}
\title{getbien}
\usage{
getbien(synonyms_list)
}
\arguments{
\item{synonyms_list}{A list of affiliated names for your query.}
}
\value{
a data frame with desired columns from BIEN.
}
\description{
This function queries BIEN for your desired species.
}
\details{
This function is used in the gators_download function.
This function uses the correct_class function.
This function requires the packages BIEN, magrittr, base, dplyr.
}
\keyword{internal}
