% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgbif.R
\name{getgbif}
\alias{getgbif}
\title{getgbif}
\usage{
getgbif(synonyms_list, gbif_match = "fuzzy")
}
\arguments{
\item{synonyms_list}{A list of affiliated names for your query.}

\item{gbif_match}{Either "f" for fuzzy matching of name or "s" to search by species code.}
}
\value{
Returns a data frame with desired columns from GBIF.
}
\description{
This function queries the Global Biodiversity Information Facility (GBIF) for your desired species.
Limited to 100000 record downloads.
}
\details{
This function is used in the gators_download function.
This function uses the correct_class function.
This function requires the packages rgbif, base, magrittr, dplyr.
}
