% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getidigbio.R
\name{getidigbio}
\alias{getidigbio}
\title{Download data from Integrated Digitized Biocollections}
\usage{
getidigbio(synonyms_list)
}
\arguments{
\item{synonyms_list}{A list of affiliated names for your query.}
}
\value{
A data frame with desired columns from iDigBio.
}
\description{
The \code{getidigbio()} function queries iDigBio for your desired species.
Limited to 100,000 record downloads.
}
\details{
This function uses the \code{correct_class()} function.
This function requires the packages ridigbio, magrittr, and dplyr.
}
\examples{
df <- getidigbio(c("Galax urceolata", "Galax aphylla"))

}
