% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchColClasses.R
\name{matchColClasses}
\alias{matchColClasses}
\title{matchColClasses}
\usage{
matchColClasses(df1, df2)
}
\arguments{
\item{df1}{A data frame}

\item{df2}{A data frame}
}
\value{
df2 with corrected column classes
}
\description{
This function allows you to match the column classes of two data frames.
This function requires package base. This function does not require user input.
Source: https://stackoverflow.com/questions/49215193/r-error-cant-join-on-because-of-incompatible-types
}
