% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/need_to_georeference.R
\name{need_to_georeference}
\alias{need_to_georeference}
\title{need_to_georeference}
\usage{
need_to_georeference(occurrence_records)
}
\arguments{
\item{occurrence_records}{A data frame of occurrence records.}
}
\value{
for_georeferencing is a data frame of the points that need to be georeferenced. For more information about this data frame, see `gators_download()`.
}
\description{
This function allows you to find records that are missing coordinates but have locality information.
}
\details{
This function requires packages dplyr.
This function does not require user input.
}
