% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/needed_records.R
\name{needed_records}
\alias{needed_records}
\title{needed_records

This function identifies records with flags}
\usage{
needed_records(occurrence_records)
}
\arguments{
\item{occurrence_records}{is a data frame downloaded with spocc_combine}
}
\value{
a data frame with rows that locality was flagged
}
\description{
needed_records

This function identifies records with flags
}
