% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove Duplicates - Remove records with identical event dates and coordinates}
\usage{
remove_duplicates(
  df,
  event.date = "eventDate",
  latitude = "latitude",
  longitude = "longitude",
  aggregator = "aggregator",
  id = "ID"
)
}
\arguments{
\item{df}{Data frame of occurrence records returned from \code{gators_download()}.}

\item{event.date}{Default = "eventDate". The name of the event date column in the data frame.}

\item{latitude}{Default = "latitude". The name of the latitude column in the data frame.}

\item{longitude}{Default = "longitude". The name of the longitude column in the data frame.}

\item{aggregator}{Default = "aggregator". The name of the column in the data frame that identifies the aggregator
that provided the record.}

\item{id}{Default = "ID". The name of the id column in the data frame, which contains unique IDs defined from GBIF or iDigBio.}
}
\value{
Return data frame with duplicates removed.
}
\description{
The \code{remove_duplicates()} function removes records with identical event dates and coordinate values.
Prior to utilizing this function, longitude and latitude columns should be rounded to match the
coordinate uncertainty using the \code{basic_locality_clean()} function.
}
\examples{
data <- remove_duplicates(data)

}
