% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removed_redacted.R
\name{remove_redacted}
\alias{remove_redacted}
\title{Remove Redacted Information - Remove protected or private records prior to publication}
\usage{
remove_redacted(df, aggregator = "aggregator")
}
\arguments{
\item{df}{A data frame downloaded with \code{gators_download()}.}

\item{aggregator}{Default = "aggregator". The name of the column in the data frame that identifies the aggregator
that provided the record.}
}
\value{
A data frame with redacted records removed.
}
\description{
The \code{remove_redacted()} function identifies and removes records where 'aggregator' is not equal to iDigBio or GBIF.
}
\details{
This function requires no additional packages.
}
\examples{
cleaned_data <- remove_redacted(data)

}
