% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_skewed.R
\name{remove_skewed}
\alias{remove_skewed}
\title{Locality Cleaning - Removed skewed locality}
\usage{
remove_skewed(df)
}
\arguments{
\item{df}{A data frame downloaded with \code{gators_download()}.}
}
\value{
A data frame with records remove only records for which locality was skewed.
}
\description{
The \code{remove_skewed()} function identifies and removes records where locality has been skewed.
}
