% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupduckDB.R
\name{setupduckDB}
\alias{setupduckDB}
\title{Used in gators_download() - Setting up local query}
\usage{
setupduckDB()
}
\description{
Downloading a parquet database containing recordsets that were not uploaded to GBIF by June 2026,
for the desired columns. This function sets up a local duckDB database to use in database searches.
}
