% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_correction.R
\name{spatial_correction}
\alias{spatial_correction}
\title{Spatial Correction -}
\usage{
spatial_correction(df, accepted.name = NA, thin.par = 1000, reps = 1)
}
\arguments{
\item{df}{Data frame of occurrence records.}

\item{accepted.name}{Accepted name of your species. This argument is not required
if the data frame already contains an accepted_name column.}

\item{thin.par}{Default = 1000. Distance in km to separate records.}

\item{reps}{Default = 1. Number of times to perform thinning.}
}
\value{
df is a data frame with the cleaned data.
}
\description{
The \code{spatial_correction} function performs coordinate thinning.
}
\details{
This function requires package spThin.
}
\examples{
\dontrun{
data <- spatial_correction(data, accepted.name = "Galax urceolata")
}

}
\keyword{internal}
