% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppress_output.R
\name{suppress_output}
\alias{suppress_output}
\title{Suppress print statements and messages}
\usage{
suppress_output(input)
}
\arguments{
\item{function}{}
}
\value{
output Output is the output of the passed function call.
}
\description{
The \code{suppress_output()} function suppresses print statements and messages from a function. This
does not suppress warnings or errors.
}
\details{
This function requires no additional packages.
}
\keyword{internal}
