% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_clean.R
\name{taxa_clean}
\alias{taxa_clean}
\title{Taxonomic Cleaning - Filter and resolve taxon names}
\usage{
taxa_clean(df, synonyms_list, filter = "fuzzy", accepted_name = NA)
}
\arguments{
\item{df}{Data frame with name column to be fixed.}

\item{synonyms_list}{A list of synonyms for a species.}

\item{filter}{The type of filter to be used--either "exact", "fuzzy", or "interactive".}

\item{accepted_name}{The accepted scientific name for the species.}
}
\value{
Returns data frame with filtered results.
}
\description{
The \code{taxa_clean()} function filters a data frame for relevant results, based on the scientific name given.
Filtering can be done with scripts by exact or fuzzy match. Or, for a more controlled approach, this function
provides interactive filtering by providing the user with prompts. The interactive method allows the user
to manually determine whether they wish to keep results containing certain scientific names.
}
\details{
This function requires packages dplyr and magrittr.
}
\examples{
data <- taxa_clean(data, c("Galax urceolata", "Galax aphylla"), filter = "exact")
data <- taxa_clean(data, c("Galax urceolata", "Galax aphylla"), accepted_name = "Galax urceolata")

}
