% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_locality_clean.R
\name{basic_locality_clean}
\alias{basic_locality_clean}
\title{Locality Cleaning - Remove missing and improbable coordinates}
\usage{
basic_locality_clean(df, remove.zero = TRUE, precision = 2)
}
\arguments{
\item{df}{Data frame of occurrence records returned from \code{gators_download()}.}

\item{remove.zero}{Default = "TRUE". Indicates that points at (0.00, 0.00) should be removed.}

\item{precision}{Indicates digits to round coordinates too. Coordinates should be round to match the coordinate uncertainty. Default = 2.}
}
\value{
Returns data
}
\description{
The \code{basic_locality_clean()} function cleans locality by removing missing or impossible coordinates and correcting precision.
This function requires columns named 'latitude' and 'longitude'.
}
\examples{
data <- basic_locality_clean(data)

}
