% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_columns.R
\name{check_columns}
\alias{check_columns}
\title{check_columns

This function query BISON for your desired species.
This function requires package base.}
\usage{
check_columns(spocc_name, fields)
}
\arguments{
\item{spocc_name}{is a list of columns from the query}

\item{fields}{is a list of fields}
}
\value{
a data frame with desired columns
}
\description{
check_columns

This function query BISON for your desired species.
This function requires package base.
}
