% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_columns.R
\name{check_columns}
\alias{check_columns}
\title{check_columns}
\usage{
check_columns(spocc_name, fields)
}
\arguments{
\item{spocc_name}{A list of columns from the query.}

\item{fields}{A list of fields.}
}
\value{
Returns a data frame with desired columns.
}
\description{
This function query BISON for your desired species.
This function requires package base.
}
