% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_bellow.R
\name{citation_bellow}
\alias{citation_bellow}
\title{Cite Data - Get species, longitude, and latitude columns}
\usage{
citation_bellow(df, id = "ID", aggregator = "aggregator")
}
\arguments{
\item{df}{Data frame of occurrence records returned from \code{gators_download()}.}

\item{id}{Default = "ID". The name of the id column in the data frame, which contains unique IDs defined from GBIF or iDigBio.}

\item{aggregator}{Default = "aggregator". The name of the column in the data frame that identifies the aggregator
that provided the record.}
}
\value{
Returns a list with citation information for the GBIF data downloaded.
}
\description{
The \code{citation_bellow} function retrieves and returns the citation information
for the data provided by GBIF in a data frame.
}
\details{
This function requires the rgbif package.
}
\examples{
citations <- citation_bellow(data)

}
