% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_class.R
\name{correct_class}
\alias{correct_class}
\title{correct_class}
\usage{
correct_class(df)
}
\arguments{
\item{df}{Data frame returned by `gator_download()`.}
}
\value{
Returns data frame with corrected classes of each column.
}
\description{
This function corrects the classes of each column in a data frame of your queried species.
This function requires the package base.
}
