% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Downloaded data from gators_download() for Asclepias curtissii with default settings:
gators_download(c("Asclepias curtissii", "Asclepias aceratoides", "Asclepias arenicola", "Oxypteryx arenicola", "Oxypteryx curtissii"), "newFile.csv")}
\format{
A data frame with 559 rows and 19 columns
}
\usage{
data
}
\description{
Downloaded data from gators_download() for Asclepias curtissii with default settings:
gators_download(c("Asclepias curtissii", "Asclepias aceratoides", "Asclepias arenicola", "Oxypteryx arenicola", "Oxypteryx curtissii"), "newFile.csv")
}
\keyword{internal}
