% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fix_names.R
\name{filter_fix_names}
\alias{filter_fix_names}
\title{gatoRs Download -  Filter iDigBio results by scientific name}
\usage{
filter_fix_names(df, synonyms_list, filter = "fuzzy", accepted_name)
}
\arguments{
\item{df}{Data frame with name column to be fixed.}

\item{synonyms_list}{A list of synonyms for a species.}

\item{filter}{The type of filter to be used--either "exact" or "fuzzy".}

\item{accepted_name}{The accepted scientific name for the species.}
}
\value{
Returns data frame with filtered results.
}
\description{
The \code{filter_fix_names()} function filters a data frame for relevant results, based on the scientific name given.
Some downloaded results from iDigBio might contain occurrences of other species that have "notes" mentioning
the desired species. Hence, this function looks for relevant results that are actually occurrences of the
desired species.
}
\details{
This function requires no additional packages.
}
\examples{
data \%>\% filter_fix_names(c("Galax urceolata", "Galax aphylla"), filter = "exact")
data \%>\% filter_fix_names(c("Galax urceolata", "Galax aphylla"), accepted_name = "Galax urceolata")

}
