% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fix_names.R
\name{filter_fix_names}
\alias{filter_fix_names}
\title{filter_fix_names}
\usage{
filter_fix_names(df, synonyms_list, filter = "fuzzy", accepted_name)
}
\arguments{
\item{df}{Data frame with name column to be fixed.}

\item{synonyms_list}{A list of synonyms for a species.}

\item{filter}{The type of filter to be used--either "exact" or "fuzzy".}

\item{accepted_name}{The accepted scientific name for the species.}
}
\value{
Returns data frame with filtered results.
}
\description{
This function filters a data frame for relevant results, based on the scientific name given.
}
\details{
This function requires the base package.
}
