% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_select_name.R
\name{filter_select_name}
\alias{filter_select_name}
\title{filter_select_name}
\usage{
filter_select_name(df, synonyms_list, filter, accepted_name)
}
\arguments{
\item{df}{Data frame with name column to be fixed.}

\item{synonyms_list}{A list of synonyms for a species.}

\item{filter}{The type of filter to be used--either "exact", "fuzzy", or "interactive".}

\item{accepted_name}{The accepted scientific name for the species.}
}
\value{
Returns data frame with filtered results.
}
\description{
This function filters a data frame for relevant results, based on the scientific name given.
}
\details{
This function requires packages base and dplyr.
}
