% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_clean.R
\name{final_clean}
\alias{final_clean}
\title{final_clean}
\usage{
final_clean(df, cluster = TRUE, precision = 2, recordBasis = TRUE)
}
\arguments{
\item{df}{is a dataframe of occurrence records}

\item{cluster}{is an option (TRUE/FALSE) to cluster geographically close occurrence records. This option is turned on by default.}

\item{precision}{in an option to round coordinates to the specified number of digits. This option is set to 2 by default. To
not use any rounding, choose `precision = NA`.}

\item{recordBasis}{is an option (TRUE/FALSE) to interactively remove types of basis of record. This option is turned on by default.}
}
\value{
df is a dataframe with the cleaned data
}
\description{
This function performs final cleaning steps, including: removing
duplicate data points, checking locality precision, and
retaining only one collection point.
}
\details{
This function requires packages dplyr, base, raster, spatstat
}
