% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_clean.R
\name{final_clean}
\alias{final_clean}
\title{final_clean}
\usage{
final_clean(df)
}
\arguments{
\item{df}{is a dataframe of occurrence records}
}
\value{
df is a dataframe with the cleaned data
}
\description{
This function performs final cleaning steps, including: removing
duplicate data points, checking locality precision, and
retaining only one collection point.
}
\details{
This function requires packages dplyr, base, raster, spatstat, CoordinateCleaner
}
