% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_flagged.R
\name{find_flagged}
\alias{find_flagged}
\title{Find possibly problematic occurrence records}
\usage{
find_flagged(df)
}
\arguments{
\item{df}{Data frame of occurrence records.}
}
\value{
Return a copy of the original dataframe provided as input but without any points the user chose to remove.
}
\description{
The \code{find_flagged()} function allows you to find and map possible problematic points and
manually inspect and remove these points, if desired.
}
\details{
This function requires packages dplyr, CoordinateCleaner, leaflet, and magrittr.
This function requires interactive user input.
}
