% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_columns.R
\name{fix_columns}
\alias{fix_columns}
\title{fix_columns}
\usage{
fix_columns(df)
}
\arguments{
\item{df}{Data frame of occurrence records.}
}
\value{
Returns the original data frame with the specified columns.
}
\description{
This function fills out the taxonomic name columns based on available information.
}
\details{
This function requires packages dplyr, base, magrittr, stringr.
This function uses the correct_class function.
}
