% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_columns.R
\name{fix_columns}
\alias{fix_columns}
\title{Fill out taxonomic name columns}
\usage{
fix_columns(df)
}
\arguments{
\item{df}{Data frame of occurrence records.}
}
\value{
Returns the original data frame with the specified columns.
}
\description{
The \code{fix_columns()} function fills out the taxonomic name columns based on available information in the data set.
For example, if a row has a name provided for the scientificName column, this information will be used
to generate the respective genus, specificEpithet, and infraspecificEpithet columns for that row.
}
\details{
This function requires package stringr.
}
\examples{
data \%>\% fix_columns()

}
