% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_names.R
\name{fix_names}
\alias{fix_names}
\title{gatoRs Download - Fix taxonomic name capitalization}
\usage{
fix_names(df)
}
\arguments{
\item{df}{Data frame with name column to be fixed.}
}
\value{
Returns df with fixed capitalization in name column.
}
\description{
The \code{fix_names()} function fixes the capitalization of species names in the data frame provided to align
with accepted capitalization standards.
}
\details{
This function uses the \code{fixAfterPeriod()} function. This function requires package stringr.
}
\examples{
data \%>\% fix_names()

}
