% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_names.R
\name{fix_names}
\alias{fix_names}
\title{fix_names}
\usage{
fix_names(df)
}
\arguments{
\item{df}{Data frame with name column to be fixed.}
}
\value{
Returns df with fixed capitalization in name column.
}
\description{
This function fixes the capitalization of species names.
}
\details{
This function requires packages base and stringr. This function does not require user input.
}
