% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixAfterPeriod.R
\name{fixAfterPeriod}
\alias{fixAfterPeriod}
\title{fixAfterPeriod

This function fixes capitalization of species names when there are periods in the name.
Requires packages base and stringr. Does not require user input.}
\usage{
fixAfterPeriod(substring)
}
\arguments{
\item{substring}{is a substring from the name column of the dataframe to be fixed}
}
\value{
returns the substring with fixed capitalization
}
\description{
fixAfterPeriod

This function fixes capitalization of species names when there are periods in the name.
Requires packages base and stringr. Does not require user input.
}
