% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixAfterPeriod.R
\name{fixAfterPeriod}
\alias{fixAfterPeriod}
\title{Fix taxonomic capitalization of a species name when there are periods involved}
\usage{
fixAfterPeriod(substring)
}
\arguments{
\item{substring}{A substring from the name column of the data frame to be fixed.}
}
\value{
Returns the substring with fixed capitalization.
}
\description{
The \code{fixAfterPeriod()} function fixes taxonomic capitalization of a string when there are periods in the species name.
}
\details{
Requires package stringr.
}
\keyword{internal}
