% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gators_download.R
\name{gators_download}
\alias{gators_download}
\title{Download specimen data from both iDigBio and GBIF}
\usage{
gators_download(
  synonyms.list,
  write.file = FALSE,
  filename = NA,
  gbif.match = "fuzzy",
  idigbio.filter = TRUE
)
}
\arguments{
\item{synonyms.list}{A list of synonyms for your desired species. For example, \code{synonyms.list = c("Asclepias curtissii","Asclepias aceratoides", "Asclepias arenicola", "Oxypteryx arenicola", "Oxypteryx curtissii")}.
This parameter is required.}

\item{write.file}{A parameter to choose whether to produce a .csv file containing search results.
This parameter is not required and is assigned FALSE by default.}

\item{filename}{The path and file name for the retrieved data. Note that this parameter should include the ".csv"
extension as well. For example, \code{filename = "base_folder/other_folder/my_file.csv"}. The file path can be entered
either as relative to the current working directory (example: "../my_file.csv") or as a full path. This parameter is
required if \code{write.file = TRUE}.}

\item{gbif.match}{A parameter to select either search by fuzzy matching of scientific name or to search by species code.
For example, \code{gbif.match = "fuzzy"} will search by fuzzy match and \code{gbif.match = "code"} will search by code. This parameter
is not required and is assigned "fuzzy" by default.}

\item{idigbio.filter}{A parameter to remove less relevant search results from iDigBio. Based on the search input, results may
include data points for a different species that mention the desired species in the locality information, for example.
Choosing \code{idigbio.filter = TRUE} will return the data frame with rows in which the name column fuzzy matches a name on the synonym list.
This parameter is not required and is assigned TRUE by default.}
}
\value{
Returns a data frame and writes a csv file as specified in the input.
This csv file will contain search results for the desired species
from the GBIF and iDigBio databases. The columns are as follows:
\itemize{
\item \href{http://rs.tdwg.org/dwc/terms/scientificName}{scientificName}
\item \href{https://dwc.tdwg.org/list/#dwc_genus}{genus}
\item \href{https://dwc.tdwg.org/list/#dwc_specificEpithet}{specificEpithet}
\item \href{http://rs.tdwg.org/dwc/terms/infraspecificEpithet}{infraspecificEpithet}
\item \href{http://rs.tdwg.org/dwc/terms/basisOfRecord}{basisOfRecord}
\item \href{http://rs.tdwg.org/dwc/terms/eventDate}{eventDate}
\item \href{http://rs.tdwg.org/dwc/terms/institutionCode}{institutionCode}
\item \href{http://rs.tdwg.org/dwc/terms/collectionCode}{collectionCode}
\item \href{http://rs.tdwg.org/dwc/terms/collectionID}{collectionID}
\item \href{http://rs.tdwg.org/dwc/terms/identificationID}{identificationID}
\item \href{http://rs.tdwg.org/dwc/terms/informationWithheld}{informationWithheld}
\item \href{http://rs.tdwg.org/dwc/terms/country}{country}
\item \href{http://rs.tdwg.org/dwc/terms/county}{county}
\item \href{http://rs.tdwg.org/dwc/terms/stateProvince}{stateProvince}
\item \href{http://rs.tdwg.org/dwc/terms/locality}{locality}
\item \href{http://rs.tdwg.org/dwc/terms/decimalLatitude}{latitude}
\item \href{http://rs.tdwg.org/dwc/terms/decimalLongitude}{longitude}
\item \href{http://rs.tdwg.org/dwc/terms/coordinateUncertaintyInMeters}{coordinateUncertaintyInMeters}
\item \href{http://rs.tdwg.org/dwc/terms/habitat}{habitat}
}
}
\description{
The \code{gators_download()} function downloads data from GBIF and iDigBio for your desired species. Data returned includes:
scientificName, genus, specificEpithet, infraspecificEpithet, basisOfRecord, eventDate, institutionCode, collectionCode,
collectionID, country, county, stateProvince, locality, latitude, longitude, identificationID,
coordinateUncertaintyInMeters, informationWithheld, and habitat.
}
\details{
This function uses the \code{get_idigbio()}, \code{get_gbif()}, \code{fix_columns()}, \code{fix_names()}, and \code{filter_fix_names()} functions.
This function requires packages magrittr, rgbif, dplyr, ridigbio, and stringr.
}
\examples{
df <- gators_download(synonyms.list = c("Galax urceolata", "Galax aphylla"), write.file = TRUE, filename = "galax.csv")
df <- gators_download(synonyms.list = "Galax urceolata", gbif.match = "code", idigbio.filter = FALSE)

}
