% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_idigbio.R
\name{get_idigbio}
\alias{get_idigbio}
\title{Download data from Integrated Digitized Biocollections}
\usage{
get_idigbio(synonyms.list)
}
\arguments{
\item{synonyms.list}{A list of affiliated names for your query.}
}
\value{
A data frame with desired columns from iDigBio.
}
\description{
The \code{get_idigbio()} function queries iDigBio for your desired species.
Limited to 100,000 record downloads.
}
\details{
This function uses the \code{correct_class()} function.
This function requires the packages ridigbio, magrittr, and dplyr.
}
\examples{
df <- get_idigbio(c("Galax urceolata", "Galax aphylla"))

}
