% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgbif.R
\name{getgbif}
\alias{getgbif}
\title{Download data from the Global Biodiversity Information Facility}
\usage{
getgbif(synonyms_list, gbif_match = "fuzzy")
}
\arguments{
\item{synonyms_list}{A list of affiliated names for your query.}

\item{gbif_match}{Either "fuzzy" for fuzzy matching of name or "code" to search by species code.}
}
\value{
Returns a data frame with desired columns from GBIF.
}
\description{
The \code{getgbif()} function queries the Global Biodiversity Information Facility (GBIF) for your desired species.
Limited to 100,000 record downloads.
}
\details{
This function uses the \code{correct_class()} function.
This function requires the packages rgbif, magrittr, and dplyr.
}
