% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgbif.R
\name{getgbif}
\alias{getgbif}
\title{getgbif

This function queries GBIF for your desired species.
This function is used in the gators_download function.
This function uses the correct_class function.
This function requires the packages rgbif, base, magrittr, dplyr.}
\usage{
getgbif(synonyms_list, gbif_match = "fuzzy")
}
\arguments{
\item{synonyms_list}{is a list of affiliated names for your query.}

\item{gbif_match}{is either "f" for fuzzy matching of name or "s" to search by species code.}
}
\value{
a data frame with desired columns from GBIF.
Limited to 100000 record downloads.
}
\description{
getgbif

This function queries GBIF for your desired species.
This function is used in the gators_download function.
This function uses the correct_class function.
This function requires the packages rgbif, base, magrittr, dplyr.
}
