% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getidigbio.R
\name{getidigbio}
\alias{getidigbio}
\title{getidigbio

This function queries iDigBio for your desired species.
This function is used in the gators_download function.
This function uses the correct_class function.
This function requires the packages ridigbio, magrittr, base, dplyr.}
\usage{
getidigbio(synonyms_list)
}
\arguments{
\item{synonyms_list}{is a list of affiliated names for your query.}
}
\value{
a data frame with desired columns from iDigBio.
Limited to 100000 record downloads.
}
\description{
getidigbio

This function queries iDigBio for your desired species.
This function is used in the gators_download function.
This function uses the correct_class function.
This function requires the packages ridigbio, magrittr, base, dplyr.
}
