% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getidigbio.R
\name{getidigbio}
\alias{getidigbio}
\title{getidigbio}
\usage{
getidigbio(synonyms_list)
}
\arguments{
\item{synonyms_list}{A list of affiliated names for your query.}
}
\value{
A data frame with desired columns from iDigBio.
}
\description{
This function queries iDigBio for your desired species.
Limited to 100000 record downloads.
}
\details{
This function is used in the gators_download function.
This function uses the correct_class function.
This function requires the packages ridigbio, magrittr, base, dplyr.
}
