% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/need_to_georeference.R
\name{need_to_georeference}
\alias{need_to_georeference}
\title{need_to_georeference}
\usage{
need_to_georeference(occurrence_records)
}
\arguments{
\item{occurrence_records}{is a dataframe of occurrence records}
}
\value{
for_georeferencing_all is a dataframe of the points that need to be georeferenced
}
\description{
This function allows you to find the points that are missing coordinates
but have a locality.
}
\details{
This function requires packages dplyr, base, magrittr.
This function does not require user input.
}
