% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/need_to_georeference.R
\name{need_to_georeference}
\alias{need_to_georeference}
\title{Identify Missing Information - Find records which lack coordinate information}
\usage{
need_to_georeference(
  df,
  longitude = "longitude",
  latitude = "latitude",
  locality = "locality"
)
}
\arguments{
\item{df}{A data frame downloaded with \code{gators_download()}.}

\item{longitude}{Default = "longitude". The name of the \href{http://rs.tdwg.org/dwc/terms/decimalLongitude}{longitude} column in the data frame.}

\item{latitude}{Default = "latitude". The name of the \href{http://rs.tdwg.org/dwc/terms/decimalLatitude}{latitude} column in the data frame.}

\item{locality}{Default = "locality". The name of the \href{http://rs.tdwg.org/dwc/terms/locality}{locality} column in the data frame.}
}
\value{
Returns a data frame of the points that need to be georeferenced.
For more information about this data frame, see \code{gators_download()}.
}
\description{
The \code{need_to_georeference()} function allows you to find records that are missing coordinates
but contain locality information. These records can then be manually georeferenced.
}
\details{
This function requires no additional packages.
}
\examples{
need_coords <- need_to_georeference(data)

}
