% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/needed_records.R
\name{needed_records}
\alias{needed_records}
\title{Find records with redacted or missing data}
\usage{
needed_records(occurrence_records)
}
\arguments{
\item{occurrence_records}{A data frame downloaded with \code{gators_download()}.}
}
\value{
a data frame with rows that locality was flagged
}
\description{
The \code{needed_records()} function identifies records with flags. This indicates that information
is withheld from these records due to endangered species status, for example. Accessing this information may
require a permit. Or, these records can be removed from the data set.
}
\details{
This function requires packages dplyr, and magrittr.
}
