% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/needed_records.R
\name{needed_records}
\alias{needed_records}
\title{needed_records}
\usage{
needed_records(occurrence_records)
}
\arguments{
\item{occurrence_records}{A data frame downloaded with gators_download.}
}
\value{
a data frame with rows that locality was flagged
}
\description{
This function identifies records with flags
}
\details{
This function requires packages dplyr, magrittr, base.
}
