% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_point_per_pixel.R
\name{one_point_per_pixel}
\alias{one_point_per_pixel}
\title{Spatial Correction - One point per pixel}
\usage{
one_point_per_pixel(df, raster = NA, resolution = 0.5)
}
\arguments{
\item{df}{Data frame of occurrence records.}

\item{raster}{Raster object which will be used for ecological niche comparisons.}

\item{resolution}{Default = 0.5. Options - 0.5, 2.5, 5, and 10 (in min of a degree). 0.5 min of a degree is equal to 30 arc sec.}
}
\value{
df is a data frame with only one point per pixel.
}
\description{
The \code{one_point_per_pixel} retains only one point per pixel.
}
\details{
This function requires package raster and spatstat.geom.
}
