% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove Duplicates - Remove records with identical event dates and coordinates}
\usage{
remove_duplicates(df)
}
\arguments{
\item{df}{Data frame of occurrence records returned from \code{gators_download()}.}
}
\value{
Return data frame with duplicates removed.
}
\description{
The \code{remove_duplicates()} function removes records with identical event dates and coordinate values.
Prior to utilizing this function, longitude and latitude columns should be rounded to match the coordinate uncertainty using the \code{basic_locality_clean()} function.
}
\examples{
data <- remove_duplicates(data)

}
