% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removed_redacted.R
\name{remove_redacted}
\alias{remove_redacted}
\title{Remove Missing Information - Remove records with redacted or missing data}
\usage{
remove_redacted(df, info.withheld = "informationWithheld")
}
\arguments{
\item{df}{A data frame downloaded with \code{gators_download()}.}

\item{info.withheld}{Default = "informationWithheld". The name of the information withheld column in the data frame.}
}
\value{
A data frame with redacted records removed.
}
\description{
The \code{remove_redacted()} function identifies and removes records with flags. This indicates that information
is withheld from these records due to endangered species status, for example. Accessing this information may
require a permit.
}
\details{
This function requires no additional packages.
}
\examples{
cleaned_data <- remove_redacted(data)

}
