% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_skewed.R
\name{remove_skewed}
\alias{remove_skewed}
\title{Used in basic_locality_clean() - Removed skewed locality}
\usage{
remove_skewed(df, info.withheld = "informationWithheld")
}
\arguments{
\item{df}{A data frame downloaded with \code{gators_download()}.}

\item{info.withheld}{Default = "informationWithheld". The name of the information withheld column in the data frame.}
}
\value{
A data frame with records remove only records for which locality was skewed.
}
\description{
The \code{remove_skewed()} function identifies and removes records where locality has been skewed.
}
\examples{
data <- remove_skewed(data)

}
