% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_clean.R
\name{taxa_clean}
\alias{taxa_clean}
\title{Taxonomic Cleaning - Filter and resolve taxon names}
\usage{
taxa_clean(
  df,
  synonyms.list,
  taxa.filter = "fuzzy",
  scientific.name = "scientificName",
  accepted.name = NA
)
}
\arguments{
\item{df}{Data frame of occurrence records returned from \code{gators_download()}.}

\item{synonyms.list}{A list of synonyms for a species.}

\item{taxa.filter}{The type of filter to be used--either "exact", "fuzzy", or "interactive".}

\item{scientific.name}{Default = "scientificName". The name of the scientific name column in the data frame.}

\item{accepted.name}{The accepted scientific name for the species. If provided, an additional column will be added to the data frame with the accepted name for further manual comparison.}
}
\value{
Returns data frame with filtered results and new column with the accepted name labeled as "accepted_name".
}
\description{
The \code{taxa_clean()} function filters a data frame for relevant results, based on the scientific name given.
Filtering can be done with scripts by exact or fuzzy match. Or, for a more controlled approach, this function
provides interactive filtering by providing the user with prompts. The interactive method allows the user
to manually determine whether they wish to keep results containing certain scientific names.
}
\details{
This function requires no additional packages.
}
\examples{
clean_data <- taxa_clean(data, c("Galax urceolata", "Galax aphylla"), taxa.filter = "exact")
clean_data <- taxa_clean(data, c("Galax urceolata", "Galax aphylla"),
accepted.name = "Galax urceolata")

}
